LayoutTests/platform/qt/http/tests/security/contentSecurityPolicy/xsl-allowed-expected.txt

CONSOLE MESSAGE: line 8: <html xmlns='http://www.w3.org/1999/xhtml/'><body><p>Running an XSL-T 1.0 stylesheet with a 2.0 processor.</p></body></html>
The text below should indicate that the style sheet was applied.
Style sheet applied.
